move some funcs
